# vectorize