Slight tweak to the conversion script to (partially) vectorize, although the outer event loop is still there.

Nonetheless, the speedup is x13.

For 1000 events:
```bash
python convertNanoToHDF5_L1triggerToDeepMET.py -i data/perfNano_TTbar_PU200.110X_set0.root -o data/perfNano_TTbar_PU200.110X_set0.h5 -N 1000
```

master branch: 1000 events in ~1 minute, 11 seconds, ~14 events / second
```
100%|███████████████████████████████████████████████████████████████████████| 1000/1000 [01:11<00:00, 13.93it/s]
```

vectorize branch: 1000 events in ~5 seconds, ~182 events / second
```
100%|██████████████████████████████████████████████████████████████████████| 1000/1000 [00:05<00:00, 182.47it/s]
```